+++ src/app/containers/Profile/CreateAddressLabel.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'react-i18next';
 import { translations } from 'locales/i18n';
 import { Form, Modal, Input, message } from '@cfxjs/antd';
 import { publishRequestError } from 'utils';
-import { LOCALSTORAGE_KEYS_MAP } from 'utils/constants';
 import { useGlobalData } from 'utils/hooks/useGlobal';
 import SDK from 'js-conflux-sdk/dist/js-conflux-sdk.umd.min.js';
+import { LOCALSTORAGE_KEYS_MAP } from 'utils/enum';
 
 type Type = {
   a: string;
