+++ src/app/containers/Profile/CreateTxNote.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next';
 import { translations } from 'locales/i18n';
 import { Form, Modal, Input, message } from '@cfxjs/antd';
 import { isHash, publishRequestError } from 'utils';
-import { LOCALSTORAGE_KEYS_MAP } from 'utils/constants';
 import { useGlobalData } from 'utils/hooks/useGlobal';
+import { LOCALSTORAGE_KEYS_MAP } from 'utils/enum';
 
 type Type = {
   h: string;
