+++ src/utils/constants.ts
@@ -1,6 +1,7 @@
 import SDK from 'js-conflux-sdk/dist/js-conflux-sdk.umd.min.js';
 import lodash from 'lodash';
 import ENV_CONFIG, { DOMAIN, IS_DEVNET, IS_STAGE } from 'env';
+import { LOCALSTORAGE_KEYS_MAP } from './enum';
 
 interface ContractsType {
   faucet: string;
@@ -18,28 +19,6 @@ interface ContractsType {
 
 interface ContractNameTagType {
   [index: string]: string;
-}
-
-/**
- * @todo
- * 1. setNFTCacheInfo cacheKey
- * 2. GlobalTip
- *
- * @export
- * @enum {number}
- */
-export enum LOCALSTORAGE_KEYS_MAP {
-  reqProjectConfigMD5 = 'CONFLUX_SCAN_REQ_PROJECT_CONFIG_MD5',
-  networkId = 'CONFLUX_SCAN_NETWORK_ID',
-  contracts = 'CONFLUX_SCAN_CONTRACTS',
-  contractNameTag = 'CONFLUX_SCAN_CONTRACT_NAME_TAG',
-  currency = 'CONFLUX_SCAN_LOCALSTORAGE_KEY_CURRENCY',
-  ageFormat = 'CONFLUX_SCAN_TABLE_AGE_FORMAT',
-  cookieAgreed = 'CONFLUXSCAN_COOKIE_AGREED',
-  txnRecords = 'CONFLUXSCAN_TXN_RECORDS',
-  fccfxNotice = 'CONFLUX_SCAN_FCCFX_NOTICE',
-  addressLabel = 'CONFLUX_SCAN_ADDRESS_LABELS',
-  txPrivateNote = 'CONFLUX_SCAN_TX_PRIVATE_NOTES',
 }
 
 export const NETWORK_ID = (() => {
@@ -179,7 +158,7 @@ export const OPEN_API_URLS = Object.entries({
   NFTBalance: '/nft/balances',
 })
   .map(item => ({
-    [item[0]]: `https://${
+    [item[0]]: `${
       item[1] === '/statistics/mining'
         ? ENV_CONFIG.ENV_CORE_API_HOST
         : ENV_CONFIG.ENV_API_HOST
