+++ src/utils/enum.ts
@@ -1,0 +1,22 @@
+/**
+ * @todo
+ * 1. setNFTCacheInfo cacheKey
+ * 2. GlobalTip
+ *
+ * @export
+ * @enum {number}
+ */
+export enum LOCALSTORAGE_KEYS_MAP {
+  reqProjectConfigMD5 = 'CONFLUX_SCAN_REQ_PROJECT_CONFIG_MD5',
+  networkId = 'CONFLUX_SCAN_NETWORK_ID',
+  contracts = 'CONFLUX_SCAN_CONTRACTS',
+  contractNameTag = 'CONFLUX_SCAN_CONTRACT_NAME_TAG',
+  currency = 'CONFLUX_SCAN_LOCALSTORAGE_KEY_CURRENCY',
+  ageFormat = 'CONFLUX_SCAN_TABLE_AGE_FORMAT',
+  cookieAgreed = 'CONFLUXSCAN_COOKIE_AGREED',
+  txnRecords = 'CONFLUXSCAN_TXN_RECORDS',
+  fccfxNotice = 'CONFLUX_SCAN_FCCFX_NOTICE',
+  addressLabel = 'CONFLUX_SCAN_ADDRESS_LABELS',
+  txPrivateNote = 'CONFLUX_SCAN_TX_PRIVATE_NOTES',
+  apis = 'CONFLUX_SCAN_APIS',
+}
