+++ src/utils/hooks/useTxnHistory.ts
@@ -2,7 +2,7 @@ import React, { useEffect, createContext, useContext } from 'react';
 import { useGetTxnStatus } from './useGetTxnStatus';
 import { usePortal } from './usePortal';
 import pubsub from '../pubsub';
-import { LOCALSTORAGE_KEYS_MAP } from 'utils/constants';
+import { LOCALSTORAGE_KEYS_MAP } from 'utils/enum';
 
 const noop = () => {};
